Add main page

+++ src/main/java/com/example/nienluan/repository/CategoryRepository.java
@@ -4,6 +4,6 @@ package com.example.nienluan.repository;
 import com.example.nienluan.models.Category;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CategoryRepository extends JpaRepository<Category,Long> {
+public interface CategoryRepository extends JpaRepository<Category,Integer> {
 
 }
