+++ src/main/java/com/example/nienluan/repository/PaymentmethodRepository.java
@@ -4,5 +4,5 @@ package com.example.nienluan.repository;
 import com.example.nienluan.models.Paymentmethod;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PaymentmethodRepository extends JpaRepository<Paymentmethod,Long> {
+public interface PaymentmethodRepository extends JpaRepository<Paymentmethod,Integer> {
 }
