+++ src/main/java/com/example/nienluan/repository/RoleRepository.java
@@ -4,5 +4,5 @@ import com.example.nienluan.models.Role;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.repository.PagingAndSortingRepository;
 
-public interface RoleRepository extends JpaRepository<Role, Long>, PagingAndSortingRepository<Role, Long> {
+public interface RoleRepository extends JpaRepository<Role, Integer>, PagingAndSortingRepository<Role, Integer> {
 }
